Cross-platform: add Windows & macOS builds to Beta 1 release
- Fix macOS compilation: guard sys/sysinfo.h (Linux-only) behind #ifndef __APPLE__
- Fix Clang size_t/unsigned long -> Variant ambiguity: cast to int64_t in debugger
- Change goto_frame(size_t) -> goto_frame(int64_t) for Godot binding compatibility
- Disable computed goto on Clang (strict indirect-goto analysis rejects scoped vars)
- Fix .gdextension: correct macOS framework paths (lib prefix), add macos.editor entry
- All 9 binaries built: Linux/Windows/macOS × editor/release/debug
- Release archive: VisualGasic-v3.2.0-beta1.tar.gz (486 MB, 3 platforms)

+++ src/visual_gasic_debugger.h
@@ -140,7 +140,7 @@ public:
                               int line_number, const Dictionary& variables);
     bool step_backward();
     bool step_forward();
-    bool goto_frame(size_t frame_index);
+    bool goto_frame(int64_t frame_index);
     ExecutionFrame get_current_frame() const;
     Array get_execution_history(int max_frames = 100) const;
